Completed making equivalent methods for new Element data type
I finished creating the methods. However, I can't successfully assign a
pointer of a derived data type without a compilation error. I googled
the issue and it might be a bug with gfortran.

Even this test code posted below won't compile because of the pointer
assignment in the last line of the program:

program TestPointers
	    type element
        	double precision, dimension(2):: X
        	integer, dimension(2)::xLoc
        	double precision:: u, v, p, Fn, Gn, Q
        	type (element), pointer:: N, S, E, W
    	end type element
    	type(element) :: middleTest
    	type(element) :: eastTest
    	middleTest%E=>eastTest
end program TestPointers